SystemUI: split PowerNotificationWarnings into Google-specific class
Place the DI stuff / annotations into PowerModule.

An alternate way of doing this is to just modify the PowerNotificationWarnings class directly and
avoid all the DI changes. In terms of registering for broadcasts, the only fundamental difference
between how PowerNotificationWarnings and PowerNotificationWarningsGoogleImpl operate is that the
latter uses a BroadcastDispatcher from SystemUI while the former just registers the
BroadcastReceiver directly with the Context.

PowerNotificationWarningsGoogleImpl extends PowerNotificationWarnings, so the original
BroadcastReceiver will still be registered.