[temporary] SystemUI: add mBatteryDead to KeyguardIndicationController
In the stock OS image for July 2025, a boolean mBatteryDead field was added to
com.android.systemui.statusbar.KeyguardIndicationController. There are no other changes to this
AOSP class; everything else involving mBatteryDead is done in the Google-specific
com.google.android.systemui.statusbar.KeyguardIndicationControllerGoogle class, and we handle in a
subsequent commit.

It's likely this will conflict with the July 2025 tags if they ever get pushed, since this is a
modification to an AOSP class, hence the [temporary] tag. This commit should be dropped when the
tags get pushed and this field is added to the class upstream.