fix cached notification privacy settings not updating
The `SparseBooleanArray`s mLockScreenShowNotifications and mLockScreenPrivateNotifications are only
updated in two methods each: a getter (`canShowPrivateNotificationsOnLockScreen` and
`canShowNotificationsOnLockscreen`) and an updater.

In getter methods ,
their old implementation specifies a default value in `SparseBooleanArray#get`, but this will not
actually insert any keys into the `SparseBooleanArray`. The updater methods only update values for
keys that are already inside of the array. As a result, there are never any keys inside
`mLockScreenShowNotifications` and `mLockScreenPrivateNotifications`; they both always return their
default values in the getter, which are both set to `true`.

Compare the old implementations of the getters `canShowPrivateNotificationsOnLockScreen` and
`canShowNotificationsOnLockscreen` with other upstream methods such as `bubblesEnabled`, where it
explicitly adds a new key-value into the array if the `userId` is missing.

Currently these are only used for notification ranking / setting package visibility override (seems
to be observed in com.android.systemui.statusbar.notification.interruption.CommonVisualInterruptionSuppressors.kt
and in refactors for our censored forwarding notifications feature.